# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "CGM Contact Manager Service",
    "description": "REST api for management of contacts",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4000",
  "basePath": "/",
  "tags": [
    {
      "name": "Contacts",
      "description": "API for contacts"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/contacts": {
      "post": {
        "tags": [
          "Contacts"
        ],
        "description": "Create new contact",
        "parameters": [
          {
            "name": "contact",
            "in": "body",
            "description": "Contact details",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New contact is created",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get all contacts",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Contacts"
            }
          }
        }
      },
      "put": {
        "summary": "Update contact with given id",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "name": "contact",
            "in": "body",
            "description": "Contact with new values of properties",

            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact is updated",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        }
      }

    },
    "/contacts/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "id of contact that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get contact with given id",
        "responses": {
          "200": {
            "description": "Contact is found",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete contact with given id",
        "tags": [
          "Contacts"
        ],
        "responses": {
          "200": {
            "description": "Contact is deleted",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "required": [
        "email",
        "name",
        "phone"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "readOnly": true
        },
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "name": {
          "type": "string",
          "uniqueItems": true
        },
        "phone": {
          "type": "string",
          "uniqueItems": true
        }
      }
    },
    "Contacts": {
      "type": "array",
      "$ref": "#/definitions/Contact"
    }
  }
}